Move .ackrc to XDG_CONFIG_HOME
ack requires that we set ACKRC to find this :(

Fortunately, if the file *doesn't* move, this setting is idempotent and
does nothing, because ack checks the default location if the $ACKRC file
is missing.

For user-level settings, the first file found is used.

But, there are also other ackrc files. They're read in this order
* The global, in /etc/ackrc
* The user, in $ACKRC, or ~/.ackrc if that fails
* The project, the first [_.]ackrc found while traversing ../....
* The cli's, specified via --ackrc

CLI arguments are effectively a final ackrc at the end of this list. The
last-applied option takes precedence over ones specified earlier.