Move .vimrc to .vim/ (vim doesn't respect XDG*)
This is the "2nd user vimrc file" per `vim --version`.

It's the best I can do short of hijacking the $VIM environment variable
or the convoluted solutions found at
https://wiki.archlinux.org/title/XDG_Base_Directory.

Vim is reluctant to add 'yet another' config location, which is fine,
but frustrating. This at least reduces the footprint to one $HOME entry.