Allow git to find its mailmap file in XDG_CONFIG_HOME
I only really use this at work, but it's nice to keep it with the other
git tech.

Git doesn't search for this file outside of $GIT_DIR, so we already
needed to specify its location for the affected repositories - this is
just a change of where this already-configured file lives.